% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{mark}
\alias{mark}
\alias{mark_html}
\alias{mark_latex}
\title{Render Markdown to an output format}
\usage{
mark(
  file = NULL,
  output = NULL,
  text = NULL,
  format = c("html", "latex"),
  options = NULL,
  template = FALSE,
  meta = list()
)

mark_html(..., template = TRUE)

mark_latex(..., template = TRUE)
}
\arguments{
\item{file, output, text, options, meta}{Passed to \code{\link[litedown:mark]{litedown::mark()}}.}

\item{format}{Output format name.}

\item{template}{Whether to use a built-in template, or path to a custom
template.}

\item{...}{Arguments to be passed to \code{mark()}.}
}
\description{
This is a wrapper function based on \code{\link[litedown:mark]{litedown::mark()}}. You should use
\code{litedown::mark()} directly.
}
