\name{markdown-package}
\alias{markdown-package}
\alias{markdown}
\docType{package}
\title{
Markdown to HTML renderer in R
}
\description{

}
\details{
\tabular{ll}{
Package: \tab markdown\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-03-31\cr
License: \tab What license is it under?\cr
}
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
