% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-package.R
\docType{package}
\name{markdown}
\alias{markdown}
\title{Markdown rendering for R}
\description{
\pkg{Markdown} is a plain-text formatting syntax that can be converted to
XHTML or other formats. This package provides wrapper functions based on the
\pkg{commonmark} package.
}
\details{
The R function \code{\link{markdownToHTML}} renders a markdown file to HTML
(respecting the specified \code{\link{markdownExtensions}} and
\code{\link{markdownHTMLOptions}}).

To learn more about markdown syntax see:
\url{https://en.wikipedia.org/wiki/Markdown}
}
\seealso{
\code{\link{markdownToHTML}} \code{\link{renderMarkdown}}
}
\author{
JJ Allaire, Jeffrey Horner, Vicent Marti, and Natacha Porte

  Maintainer: Yihui Xie <xie@yihui.name>
}
\keyword{package}
