% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{markdown_options}
\alias{markdown_options}
\title{Markdown rendering options}
\usage{
markdown_options()
}
\description{
A wrapper function of \code{\link[litedown:markdown_options]{litedown::markdown_options()}}.
}
