% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderMarkdown.R
\name{markdownExtensions}
\alias{markdownExtensions}
\title{Deprecated}
\usage{
markdownExtensions(...)
}
\description{
Please specify extensions via the \code{options} argument instead.
}
\keyword{internal}
