% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderMarkdown.R
\name{markdownToHTML}
\alias{markdownToHTML}
\title{Render Markdown to HTML}
\usage{
markdownToHTML(
  file = NULL,
  output = NULL,
  text = NULL,
  options = NULL,
  title = "",
  css = NULL,
  header = NULL,
  template = NULL,
  ...
)
}
\arguments{
\item{file}{Path to an input file. If not provided, it is presumed that the
\code{text} argument will be used instead. This argument can also take a
character vector of Markdown text directly. To avoid ambiguity in the
latter case, a single character string input will be treated as a file only
when the file exists or it has a file extension. If a string happens to
have a \dQuote{file extension} and should be treated as Markdown text
instead, wrap it in \code{I()}.}

\item{output}{Output file path. If not character, the results will be
returned as a character vector.}

\item{text}{A character vector of the Markdown text. By default, it is read
from \code{file}.}

\item{options}{Options to be passed to the renderer. See
\code{\link{markdownOptions}()} for all possible options. This argument can
take either a character vector of the form \code{"+option1
option2-option3"} (use \code{+} or a space to enable an option, and
\code{-} to disable an option), or a list of the form \code{list(option1 =
value1, option2 = value2, ...)}. A string \code{"+option1"} is equivalent
to \code{list(option1 = TRUE)}, and \code{"-option2"} means
\code{list(option2 = FALSE)}. Options that do not take logical values must
be specified via a list, e.g., \code{list(width = 30)}.}

\item{title}{The HTML title.}

\item{css}{Either a vector of CSS code or a file containing CSS to be
included in the output. The default value is
\code{getOption('markdown.html.css', markdown:::pkg_file('resources',
'markdown.css'))}, i.e., it can be set via the global option
\code{markdown.html.css}.}

\item{header}{Either a vector of HTML code or a file containing HTML to be
included in the header of the output. The default value is
\code{getOption('markdown.html.header')}.}

\item{template}{An HTML template file. The default value is
\code{getOption('markdown.html.template', markdown:::pkg_file('resources',
'markdown.html'))}.}

\item{...}{Unused but for backward-compatibility with previous versions of
\pkg{markdown}.}
}
\value{
Invisible \code{NULL} when output is to a file, and a character
  vector otherwise.
}
\description{
Render Markdown to HTML with optional customization (e.g., custom stylesheets
or template).
}
\examples{
markdownToHTML("Hello _World_!", options = "-standalone")
# write HTML to an output file
markdownToHTML("_Hello_, **World**!", output = tempfile())
}
\seealso{
\code{\link{renderMarkdown}()}
}
