\name{rendererExists}
\alias{rendererExists}
\title{Testing for existence of a markdown renderer}
\description{
   \code{rendererExists} determines whether or not a certain renderer exists
   in the markdown library.
}
\usage{
rendererExists(name)
}
\arguments{
   \item{name}{name of renderer.}
}
\value{
   \code{TRUE} or \code{FALSE} for whether or not the renderer exists.
}
\examples{

rendererExists("HTML")

}
\keyword{programming}
