% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{markdown-package}
\alias{markdown}
\alias{markdown-package}
\title{Markdown rendering for R}
\description{
\pkg{Markdown} is a plain-text formatting syntax that can be converted to
XHTML or other formats. This package provides wrapper functions (mainly
\code{\link[=mark]{mark()}}) based on the \pkg{commonmark} package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/markdown}
  \item Report bugs at \url{https://github.com/rstudio/markdown/issues}
}

}
\author{
\strong{Maintainer}: Yihui Xie \email{xie@yihui.name} (\href{https://orcid.org/0000-0003-0645-5666}{ORCID})

Authors:
\itemize{
  \item JJ Allaire
  \item Jeffrey Horner
}

Other contributors:
\itemize{
  \item Henrik Bengtsson [contributor]
  \item Jim Hester [contributor]
  \item Yixuan Qiu [contributor]
  \item Kohske Takahashi [contributor]
  \item Adam November [contributor]
  \item Nacho Caballero [contributor]
  \item Jeroen Ooms [contributor]
  \item Thomas Leeper [contributor]
  \item Joe Cheng [contributor]
  \item Andrzej Oles [contributor]
  \item Posit Software, PBC [copyright holder, funder]
}

}
