\name{markdown}
\alias{markdown}
\docType{package}
\title{
Markdown to HTML renderer in R
}
\description{
   \emph{markdown} is as package for rendering markdown formatted text to
   HTML. C extensions exist as well to create other renderers.

   To learn more about \emph{markdown} syntax see:

   \url{http://daringfireball.net/projects/markdown/syntax}   
}
\author{
Copyright (C) 2009-1012 by RStudio, Inc.
}
\keyword{ package }
\seealso{
\code{\link{renderMarkdown}}
}
