% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{markdown}
\alias{markdown}
\title{Markdown rendering for R}
\description{
\pkg{Markdown} is a plain-text formatting syntax that can be converted to
XHTML or other formats. This package provides wrapper functions (mainly
\code{\link{mark}()}) based on the \pkg{commonmark} package.
}
\keyword{package}
