% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderMarkdown.R
\name{markdownToHTML}
\alias{markdownToHTML}
\title{Render Markdown to HTML}
\usage{
markdownToHTML(
  file,
  output = NULL,
  text = NULL,
  options = NULL,
  title = "",
  css = NULL,
  header = NULL,
  template = NULL,
  encoding = "UTF-8"
)
}
\arguments{
\item{file}{Path to an input file. If not provided, it is presumed that the
\code{text} argument will be used instead.}

\item{output}{Output file path. If not character, the results will be
returned as a character vector.}

\item{text}{A character vector of the Markdown text. By default, it is read
from \code{file}.}

\item{options}{A list of options to be passed to the renderer. See
\code{\link{markdownOptions}()} for all possible options.}

\item{title}{The HTML title.}

\item{css}{Either a vector of CSS code or a file containing CSS to be
included in the output. The default value is
\code{getOption('markdown.html.css', markdown:::pkg_file('resources',
'markdown.css'))}, i.e., it can be set via the global option
\code{markdown.html.css}.}

\item{header}{Either a vector of HTML code or a file containing HTML to be
included in the header of the output. The default value is
\code{getOption('markdown.html.header')}.}

\item{template}{An HTML template file. The default value is
\code{getOption('markdown.html.template', markdown:::pkg_file('resources',
'markdown.html'))}.}

\item{encoding}{Ignored (always assumes the file is encoded in UTF-8).}
}
\value{
Invisible \code{NULL} when output is to a file, and a character
  vector otherwise.
}
\description{
Render Markdown to HTML with optional customization (e.g., custom stylesheets
or template).
}
\examples{
(markdownToHTML(text = "Hello _World_!", options = "+fragment_only"))
# write HTML to an output file
markdownToHTML(text = "_Hello_, **World**!", output = tempfile())
}
\seealso{
\code{\link{renderMarkdown}()}
}
