% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderMarkdown.R
\name{rendererExists}
\alias{rendererExists}
\title{Testing for existence of a markdown renderer}
\usage{
rendererExists(name)
}
\arguments{
\item{name}{name of renderer.}
}
\value{
\code{TRUE} or \code{FALSE} for whether or not the renderer exists.
}
\description{
\code{rendererExists} determines whether or not a certain renderer exists in
the markdown library.
}
\examples{
rendererExists("HTML")
}
