% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderMarkdown.R
\name{renderMarkdown}
\alias{renderMarkdown}
\title{Render Markdown to an output format}
\usage{
renderMarkdown(
  file,
  output = NULL,
  text = NULL,
  renderer = c("html", "latex"),
  options = NULL,
  extensions = markdownExtensions()
)
}
\arguments{
\item{file}{Path to an input file. If not provided, it is presumed that the
\code{text} argument will be used instead.}

\item{output}{Output file path. If not character, the results will be
returned as a character vector.}

\item{text}{A character vector of the Markdown text. By default, it is read
from \code{file}.}

\item{renderer}{An output format supported by \pkg{commonmark}, e.g.,
\code{'html'}, \code{'man'}, and \code{'text'}, etc. See the
\code{\link[commonmark:commonmark]{markdown_*}} functions in
\pkg{commonmark}.}

\item{options}{A list of options to be passed to the renderer. See
\code{\link{markdownOptions}()} for all possible options.}

\item{extensions}{A character vector of Markdown extensions See
\code{\link{markdownExtensions}()}.}
}
\value{
Invisible \code{NULL} when output is to a file, otherwise a character
  vector.
}
\description{
Render Markdown to an output format via the \pkg{commonmark} package.
}
\examples{
renderMarkdown(text = "Hello World!")
# a few corner cases
renderMarkdown(text = character(0))
renderMarkdown(text = "")
}
\seealso{
The spec of GitHub Flavored Markdown:
  \url{https://github.github.com/gfm/}
}
