% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{renderMarkdown}
\alias{renderMarkdown}
\alias{markdownToHTML}
\title{Functions in \pkg{markdown} before \code{v1.3}.}
\usage{
renderMarkdown(...)

markdownToHTML(...)
}
\arguments{
\item{...}{Arguments to be passed to new functions.}
}
\description{
These functions are kept in this package for backward-compatibility. They
will not be removed in the foreseeable future, although we recommend that you
use their new names instead: \code{renderMarkdown()} has become
\code{\link{mark}()}, and \code{markdownToHTML()} has become
\code{\link{mark_html}()}.
}
\keyword{internal}
