% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{renderMarkdown}
\alias{renderMarkdown}
\alias{markdownToHTML}
\title{Functions in \pkg{markdown} before \code{v1.3}.}
\usage{
renderMarkdown(file = NULL, output = NULL, ...)

markdownToHTML(
  file = NULL,
  output = NULL,
  ...,
  options = getOption("markdown.HTML.options"),
  title = NULL,
  stylesheet = getOption("markdown.HTML.stylesheet"),
  header = getOption("markdown.HTML.header"),
  template = getOption("markdown.HTML.template", TRUE),
  fragment.only = FALSE,
  encoding = "UTF-8"
)
}
\arguments{
\item{file, output, ..., options, template}{Arguments to be passed to new
functions.}

\item{title, stylesheet, header}{Arguments to be passed to \verb{meta = list(title = , css = , }header-includes\verb{ = )}, which is passed to \code{\link[=mark_html]{mark_html()}}.}

\item{fragment.only}{Whether to generate a fragment or a full HTML document.}

\item{encoding}{Ignored.}
}
\description{
These functions are kept in this package for backward-compatibility. They
will not be removed in the foreseeable future, although we recommend that you
use their new names instead: \code{renderMarkdown()} has become \code{\link[=mark]{mark()}}, and
\code{markdownToHTML()} has become \code{\link[=mark_html]{mark_html()}}.
}
\keyword{internal}
