% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderMarkdown.R
\name{smartypants}
\alias{smartypants}
\title{smartypants: ASCII punctuation to HTML entities}
\usage{
smartypants(file, output, text)
}
\arguments{
\item{file}{a character string giving the pathname of the file to read from.
If it is omitted from the argument list, then it is presumed that the
\code{text} argument will be used instead.}

\item{output}{a character string giving the pathname of the file to write to.
If it is omitted, then it is presumed that the user expects the results
returned as a character string.}

\item{text}{a character vector containing the \emph{markdown} text to
transform.}
}
\value{
\code{smartypants} returns NULL invisibly when output is to a file,
  and a character string otherwise.
}
\description{
\code{smartypants} transforms plain ASCII punctuation characters into
\emph{smart} typographic punctuation HTML entities.
}
\examples{
cat(smartypants(text = "1/2 (c)\n"))
}
\seealso{
\code{\link{markdownExtensions}}, \code{\link{markdownHTMLOptions}},
  \code{\link{markdownToHTML}}.

  For a description of the original \emph{markdown} version:

  \url{https://daringfireball.net/projects/markdown/}

  The original Sundown library on github:

  \url{https://github.com/vmg/sundown}

  C stubs for writing new renders are in inst/include/markdown_rstubs.[ch].
}
